Added Implimantaion for GetWardCensusMonthlyStats

+++ src/Boxfusion.Health.His.Admissions/Services/Admissions/Dto/WardCensusResponse.cs
@@ -12,7 +12,7 @@ namespace Boxfusion.Health.His.Admissions.Services.Admissions.Dto
         public int? TotalBedAvailability { get; set; }
         public int? TotalBedInWard { get; set; }
         public decimal? BedUtilisation { get; set; }
-        public decimal? AverageBedAvailability { get; set; }
-        public decimal? AverageLengthOfStay { get; set; }
+        public double? AverageBedAvailability { get; set; }
+        public double? AverageLengthOfStay { get; set; }
     }
 }
